回复api_初略实现插入回复及解决异常

+++ src/main/java/top/codingdong/imustbbs/controller/PostController.java
@@ -23,7 +23,7 @@ public class PostController {
 
     @GetMapping("/post/{id}")
     @ApiOperation(value = "帖子详情页")
-    public String post(@PathVariable(name = "id")Integer id,
+    public String post(@PathVariable(name = "id")Long id,
                        Model model){
 
         Post post = postService.viewPostById(id);
