+++ src/main/java/top/codingdong/imustbbs/model/Post.java
@@ -9,15 +9,15 @@ import lombok.Data;
 @Data
 public class Post {
 
-    private Integer id;
+    private Long id;
     private String title;
     private String description;
     private Long createTime;
     private Long updateTime;
-    private Integer creator;
-    private Integer commentCount;
-    private Integer viewCount;
-    private Integer likeCount;
+    private Long creator;
+    private Long commentCount;
+    private Long viewCount;
+    private Long likeCount;
     private String tag;
 
     private User user;
